Dead simple login functionality. Stores info in async storage and user context.
