+++ MobileUI/screens/LoginScreen.js
@@ -1,19 +1,10 @@
-import React, {useState, useContext, createContext} from 'react';
-import {
-  Alert,
-  Text,
-  //StatusBar,
-  SafeAreaView,
-  StyleSheet,
-  View,
-  TextInput,
-} from 'react-native';
+import React, {useState, useContext} from 'react';
+import {Text, SafeAreaView, StyleSheet, View, TextInput} from 'react-native';
 import Colors from '../assets/styles/colors';
 import Button from '../assets/components/CustomButton';
-// import {LogStateContext, clearAll} from '../App';
 import LogStateContext from '../contexts/LoginState';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import LandInLogScreenContext from '../contexts/LandInLogScreen';
+import UserContext from '../contexts/User';
 
 const styles = StyleSheet.create({
   screenContainer: {
@@ -58,9 +49,10 @@ const setStringUsername = async (username, value) => {
   } catch (e) {
     // save error
     console.log('error : ' + e);
+    throw e;
   }
 
-  console.log('Set username is done.');
+  // console.log('Set username in async storage is done.');
 };
 
 const storeUserLoginInfo = async (name, password, postResult) => {
@@ -71,61 +63,141 @@ const storeUserLoginInfo = async (name, password, postResult) => {
   };
 
   await setStringUsername(name, JSON.stringify(loginInfo));
-  // let r = await getUsernameValue(name);
-  // r = JSON.parse(r);
-  // console.log('result of getting username: ' + r.password);
 };
 
-export const UsernameContext = createContext('');
+/*
+ * Clears all contents in AsyncStorage
+ */
+export const clearAll = async () => {
+  try {
+    await AsyncStorage.clear();
+  } catch (e) {
+    // clear error
+    console.log('!!!! Error with clearing! > ' + e);
+    throw e;
+  }
+
+  console.log('LoginScreen: Done clearing async storage');
+};
+
+/*
+ * Gets the value stored with given username
+ */
+const getUsernameValue = async username => {
+  try {
+    const value = await AsyncStorage.getItem(username);
+    if (value !== null) {
+      // console.log('getUsername: value=' + value);
+      return JSON.parse(value);
+    }
+  } catch (e) {
+    // read error
+    console.log('getUsername error: ' + e);
+    throw e;
+  }
+  // console.log('Done getting usesrname');
+  return null;
+};
+
+const getAllKeys = async () => {
+  let keys = [];
+  try {
+    keys = await AsyncStorage.getAllKeys();
+    if (keys !== null) {
+      console.log(keys);
+      return keys;
+    }
+  } catch (e) {
+    // read key error
+    console.log('getAllKeys error: ' + e);
+    throw e;
+  }
+
+  return keys;
+  // example console.log result:
+  // ['@MyApp_user', '@MyApp_key']
+};
+
+export async function isStorageEmpty() {
+  // console.log(
+  //   'isStorageEmpty: Checking if storage is empty. Call getAllKeys()',
+  // );
+  let k = await getAllKeys();
+  if (k.length > 0) {
+    // console.log('isStorageEmpty: No');
+    return false;
+  } else {
+    // console.log('isStorageEmpty: Yes');
+    return true;
+  }
+}
+
+// export const CheckIsTokenExpired = async () => {
+export async function checkIsTokenExpired(name) {
+  // console.log('checkIsTExp - step 2.1: Call isStorageEmpty');
+
+  let emptyStorage = await isStorageEmpty();
+  if (!emptyStorage) {
+    //There are things that exist in storage
+    // console.log(
+    //   'checkIsTExp - step 2.3: There are many!! Parse expiration and current date',
+    // );
+
+    let user = await getUsernameValue(name);
+    // console.log('checkIsTExp - step 2.4: user result=' + user);
+    var exp = Date.parse(user.expiration);
+    var d1 = new Date();
+    var d = Date.parse(d1);
+
+    // console.log(
+    //   'checkIsTExp - step 2.5: Check if current date is past expiration',
+    // );
+    if (d >= exp) {
+      // console.log('checkIsTExp - Token is expired.');
+      return true;
+    } else {
+      // console.log('checkIsTExp - Token is not expired');
+      return false;
+    }
+  } else {
+    // There is nothing stored. Need to login. return true
+    // console.log('checkIsTExp - step 2.6: nothing is stored, return true');
+    return true;
+  }
+}
 
 const LoginScreen = ({navigation}) => {
-  const [name, setName] = useState();
+  const [name, setName] = useState('');
   const [password, setPassword] = useState('');
   const [isLoading, setIsLoading] = useState(false);
   const [err, setErr] = useState('');
   const {isLoggedIn, setIsLoggedIn} = useContext(LogStateContext);
-  const landInLog = useContext(LandInLogScreenContext);
-
-  // if (getToken() !== null) {
-  //   // Meaning we have already logged in
-  //   console.log('we have already recieved a login token');
-  //   setIsLoggedIn(true);
-  // }
-  // if (isLoggedIn) {
-  //   setIsLoggedIn(false);
-  // }
-  // clearAll();
-
-  //!!!! TO DO NEXT!!!!
-  // Here, we check if user already exists and if token is expired
-  // IF user has already logged in and token is not expired, navigate to my schedule!!!!!
-  if (!landInLog) {
-    console.log('LoginScreen: landInLog is false, navigate to schedule');
-    navigation.navigate('My Schedule');
-  } else {
-    console.log('LoginScreen: landInLog is true');
-  }
+  const user = useContext(UserContext);
 
   this.nameInput = React.createRef();
   this.passwordInput = React.createRef();
 
   navigation.addListener('drawerItemPress', async () => {
     if (isLoggedIn) {
-      console.log(
-        'LoginScreen: setIsLoggedIn(false), clear inputs, clear storage',
-      );
+      // console.log(
+      //   'LoginScreen: setIsLoggedIn(false), clear inputs, clear storage, clear user context',
+      // );
       setIsLoggedIn(false);
       this.passwordInput.current.clear();
       this.nameInput.current.clear();
-      // await clearAll();
+      await clearAll();
+      user.name = '';
+      user.password = '';
+      user.token = '';
+      user.expiration = '';
     }
   });
 
   const handleLogInButton = async () => {
-    console.log('setisloading to true');
+    // console.log('setisloading to true');
     setIsLoading(true);
     try {
-      console.log('Sending Username: ' + name + ' Password: ' + password);
+      // console.log('Sending Username: ' + name + ' Password: ' + password);
       // So what needs to change is we need to send name and pass, and get the 'token' and 'expiration'.
       const response = await fetch(
         'http://ec2-52-7-65-63.compute-1.amazonaws.com/Auth/Login',
@@ -141,27 +213,29 @@ const LoginScreen = ({navigation}) => {
           }),
         },
       );
-      // if (!response.ok) {
-      //   throw new Error(`Error! status: ${response.status}`);
-      // }
 
-      console.log('await response');
+      // console.log('await response');
       const result1 = await response.json();
-      console.log(result1);
+      // console.log(result1);
 
+      // Store info in async storage
       await storeUserLoginInfo(name, password, result1);
+
+      // Store info in user context
+      user.name = name;
+      user.password = password;
+      user.token = result1.token;
+      user.expiration = result1.expiration;
 
       setIsLoading(false);
     } catch (err) {
       setErr(err.message);
       console.log('set is loading false. Send an alert for this eror: ' + err);
-      setIsLoading(false);
-      Alert.alert(err.message);
+      throw err;
     }
-    // } finally {
-    // }
+
     if (!isLoading) {
-      console.log('no longer loading, set logged in true, pull up my schedule');
+      // console.log('no longer loading, set logged in true, pull up my schedule');
       setIsLoggedIn(true);
       navigation.navigate('My Schedule');
     }
@@ -169,104 +243,36 @@ const LoginScreen = ({navigation}) => {
 
   let title = "Let's Meet";
   return (
-    <UsernameContext.Provider value={{name, setName}}>
-      <SafeAreaView style={styles.screenContainer}>
-        <View style={styles.screenContainer}>
-          <Text style={styles.defaultScreentext}>{title}</Text>
-          <Text
-            style={styles.defaultScreentext}
-            onPress={() => navigation.navigate('Registration')}>
-            No account? Click to sign up
-          </Text>
-          <TextInput
-            placeholder="User Name"
-            style={styles.userName}
-            placeholderTextColor={Colors.DD_LIGHT_GRAY}
-            onChangeText={newText => setName(newText)}
-            ref={this.nameInput}
-            autoCapitalize="none"
-          />
-          <TextInput
-            placeholder="Password"
-            style={styles.userPassword}
-            placeholderTextColor={Colors.DD_LIGHT_GRAY}
-            onChangeText={newText => setPassword(newText)}
-            ref={this.passwordInput}
-            autoCapitalize="none"
-          />
-          <Button title="Go!" onPress={() => handleLogInButton()} />
-          {isLoading && (
-            <Text style={styles.defaultScreentext}>Loading...</Text>
-          )}
-          {{err} && <Text style={styles.defaultScreentext}>{err}</Text>}
-        </View>
-      </SafeAreaView>
-    </UsernameContext.Provider>
+    <SafeAreaView style={styles.screenContainer}>
+      <View style={styles.screenContainer}>
+        <Text style={styles.defaultScreentext}>{title}</Text>
+        <Text
+          style={styles.defaultScreentext}
+          onPress={() => navigation.navigate('Registration')}>
+          No account? Click to sign up
+        </Text>
+        <TextInput
+          placeholder="User Name"
+          style={styles.userName}
+          placeholderTextColor={Colors.DD_LIGHT_GRAY}
+          onChangeText={newText => setName(newText)}
+          ref={this.nameInput}
+          autoCapitalize="none"
+        />
+        <TextInput
+          placeholder="Password"
+          style={styles.userPassword}
+          placeholderTextColor={Colors.DD_LIGHT_GRAY}
+          onChangeText={newText => setPassword(newText)}
+          ref={this.passwordInput}
+          autoCapitalize="none"
+        />
+        <Button title="Go!" onPress={() => handleLogInButton()} />
+        {isLoading && <Text style={styles.defaultScreentext}>Loading...</Text>}
+        {{err} && <Text style={styles.defaultScreentext}>{err}</Text>}
+      </View>
+    </SafeAreaView>
   );
 };
 
-// LoginScreen.defaultProps = {
-//   groupName: "My"
-// }
-
 export default LoginScreen;
-
-// import {useState} from 'react';
-
-// const App = () => {
-//   const [data, setData] = useState();
-//   const [isLoading, setIsLoading] = useState(false);
-//   const [err, setErr] = useState('');
-
-//   const handleClick = async () => {
-//     setIsLoading(true);
-//     try {
-//       const response = await fetch('https://reqres.in/api/users', {
-//         method: 'POST',
-//         body: JSON.stringify({
-//           name: 'John Smith',
-//           job: 'manager',
-//         }),
-//         headers: {
-//           'Content-Type': 'application/json',
-//           Accept: 'application/json',
-//         },
-//       });
-
-//       if (!response.ok) {
-//         throw new Error(`Error! status: ${response.status}`);
-//       }
-
-//       const result = await response.json();
-
-//       console.log('result is: ', JSON.stringify(result, null, 4));
-
-//       setData(result);
-//     } catch (err) {
-//       setErr(err.message);
-//     } finally {
-//       setIsLoading(false);
-//     }
-//   };
-
-//   console.log(data);
-
-//   return (
-//     <div>
-//       {err && <h2>{err}</h2>}
-
-//       <button onClick={handleClick}>Make request</button>
-
-//       {isLoading && <h2>Loading...</h2>}
-
-//       {data && (
-//         <div>
-//           <h2>Name: {data.name}</h2>
-//           <h2>Job: {data.job}</h2>
-//         </div>
-//       )}
-//     </div>
-//   );
-// };
-
-// export default App;
